+++ changelog.md
@@ -8,6 +8,7 @@ This project uses Semantic Versioning (2.0).
 * Add frontend test setup
 * Add angular dependency and json fixtures loading for frontend tests
 * Add an angular service for api communication
+* Use a `ProjectConfig` provider to set the `apiRoot`
 
 ## v0.1